fixed logit bias and its related converter

+++ src/main/java/es/thalesalv/chatrpg/domain/model/openai/completion/ChatCompletionRequest.java
@@ -48,5 +48,5 @@ public class ChatCompletionRequest {
     private Double frequencyPenalty;
 
     @JsonProperty("logit_bias")
-    private Map<String, Integer> logitBias;
+    private Map<String, Double> logitBias;
 }
