+++ src/main/java/es/thalesalv/chatrpg/domain/model/openai/completion/TextCompletionRequest.java
@@ -47,5 +47,5 @@ public class TextCompletionRequest {
     private Double frequencyPenalty;
 
     @JsonProperty("logit_bias")
-    private Map<String, Integer> logitBias;
+    private Map<String, Double> logitBias;
 }
